Another bunch of offences fixed
